refactor of broker for channel creation
