+++ bin/kbd
@@ -4,24 +4,25 @@
  * Kinesis Broker Daemon
  */
 
-const path = require('path');
-const program = require('caporal');
+const path = require('path')
+const program = require('caporal')
 
-const startServer = require('../broker-daemon');
+const startServer = require('../broker-daemon')
 
-const { version: CLI_VERSION } = require('../package.json');
+const { version: CLI_VERSION } = require('../package.json')
 
-const RPC_ADDRESS = process.env.RPC_ADDRESS || '0.0.0.0:27492';
-const DATA_DIR = process.env.DATA_DIR || '~/.kinesis/data';
-const ENGINE_TYPE = process.env.ENGINE_TYPE || 'lnd';
-const EXCHANGE_RPC_HOST = process.env.EXCHANGE_RPC_HOST || '0.0.0.0:28492';
+const RPC_ADDRESS = process.env.RPC_ADDRESS || '0.0.0.0:27492'
+const DATA_DIR = process.env.DATA_DIR || '~/.kinesis/data'
+const ENGINE_TYPE = process.env.ENGINE_TYPE || 'lnd'
+const EXCHANGE_RPC_HOST = process.env.EXCHANGE_RPC_HOST || '0.0.0.0:28492'
 
 // TODO: Do not provide defaults for these certs because they could
 //   exist in an array of places?
-const LND_TLS_CERT = process.env.LND_TLS_CERT || '~/.lnd/tls.cert';
-const LND_MACAROON = process.env.LND_MACAROON || '~/.lnd/admin.macaroon';
-const LND_RPC_HOST = process.env.ENGINE_RPC_HOST || '0.0.0.0:10009';
+const LND_TLS_CERT = process.env.LND_TLS_CERT || '~/.lnd/tls.cert'
+const LND_MACAROON = process.env.LND_MACAROON || '~/.lnd/admin.macaroon'
+const LND_RPC_HOST = process.env.ENGINE_RPC_HOST || '0.0.0.0:10009'
 
+// TODO: Add validations to ./bin/kbd when they become available
 program
   .version(CLI_VERSION)
   .option('--rpc-address <server>', 'Host and port to set up the RPC server on.', /^.+(:[0-9]*)?$/, RPC_ADDRESS)
@@ -34,6 +35,6 @@ program
   .option('--lnd-rpc <server>', 'Location of the LND RPC server to use.', /^.+(:[0-9]*)?$/, LND_RPC_HOST)
   .option('--lnd-tls <path>', 'Location of the certificate to use when communicating with LND.', /^.+$/, LND_TLS_CERT)
   .option('--lnd-macaroon <path>', 'Location of the macaroon to use when communicating with LND.', /^.+$/, LND_MACAROON)
-  .action(startServer);
+  .action(startServer)
 
-program.parse(process.argv);
+program.parse(process.argv)
