+++ broker-daemon/broker-actions/index.js
@@ -1,0 +1,7 @@
+const createOrder = require('./create-order')
+const watchMarket = require('./watch-market')
+
+module.exports = {
+  createOrder,
+  watchMarket
+}
