+++ broker-daemon/utils/index.js
@@ -1,0 +1,6 @@
+/**
+ * @author kinesis
+ */
+const loadProto = require('./load-proto')
+
+module.exports = { loadProto }
